-linux build fixes
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@8277 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/recon/RemoteParticipant.cxx
@@ -1031,7 +1031,7 @@ RemoteParticipant::answerMediaLine(SdpContents::Session::Medium& mediaSessionCap
               capsCodecsIt != mediaSessionCaps.codecs().end(); capsCodecsIt++)
          {
             if(isEqualNoCase(capsCodecsIt->getName(), itCodec->getMimeSubtype()) &&
-               capsCodecsIt->getRate() == itCodec->getRate())
+               (unsigned int)capsCodecsIt->getRate() == itCodec->getRate())
             {
                bool modeInCaps = capsCodecsIt->parameters().prefix("mode=");
                if(!modeInOffer && !modeInCaps)
